Use DTS-HD MA and WAVE as example in encoding workflow.
DTS-HD MA is a better example input codec as it implicitly reinforces
the rule to only transcode lossless files. The previous example used a
'.dts' file extension which is generally used for regular, lossy DTS
audio.

Also, use WAVE instead of WAV as that is the official name[1]. It also
reinforces the fact that the name of the file format doesn't have to
correspond to the file extension.

[1]: http://www-mmsp.ece.mcgill.ca/Documents/AudioFormats/WAVE/WAVE.html